import-scripts/py3_requirements.txt
certifi==2020.6.20
Unidecode==1.1.1
pandas==1.3.5
GitPython==3.1.29
python-json-logger==2.0.7
databricks-sql-connector==2.9.6